run test netcoreapp only
