+++ src/libraries/System.Security.Principal.Windows/tests/System.Security.Principal.Windows.Tests.csproj
@@ -7,6 +7,7 @@
     <Compile Include="NTAccount.cs" />
     <Compile Include="SecurityIdentifierTests.cs" />
     <Compile Include="WindowsIdentityTests.cs" />
+    <Compile Include="WindowsIdentityTests.netcoreapp.cs" Condition="'$(TargetsNetCoreApp)' == 'true'"/>
     <Compile Include="WindowsPrincipalTests.cs" />
     <Compile Include="WellKnownSidTypeTests.cs" />
   </ItemGroup>
